@@ -1,10 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 import 'package:albocarride/screens/auth/role_selection_page.dart';
+import 'package:albocarride/screens/auth/signup_page.dart';
+import 'package:albocarride/screens/auth/vehicle_type_selection_page.dart';
+import 'package:albocarride/screens/driver/verification_page.dart';
+import 'package:albocarride/screens/driver/waiting_for_review_page.dart';
 import 'package:albocarride/screens/home/customer_home_page.dart';
-import 'package:albocarride/screens/home/driver_home_page.dart';
+import 'package:albocarride/screens/home/enhanced_driver_home_page.dart';
 import 'package:albocarride/widgets/custom_toast.dart';
-import 'package:albocarride/services/session_service.dart';
 
 class AuthWrapper extends StatefulWidget {
   const AuthWrapper({super.key});
@@ -20,174 +23,191 @@ class _AuthWrapperState extends State<AuthWrapper> {
   @override
   void initState() {
     super.initState();
-    _checkAuthState();
+    _checkAndRoute();
   }
 
-  Future<void> _checkAuthState() async {
-    print('Checking authentication state...');
-    await Future.delayed(
-      const Duration(milliseconds: 500),
-    ); // Small delay for UI to show
-
-    // First check if we have a valid session stored locally
-    final hasValidSession = await SessionService.hasValidSession();
-    print('Has valid local session: $hasValidSession');
-
-    final session = _supabase.auth.currentSession;
-    print('Supabase session exists: ${session != null}');
-
-    if (session != null) {
-      await _handleAuthenticatedSession(session);
-    } else if (hasValidSession) {
-      // We have a local session but no Supabase session - try to restore
-      await _tryRestoreSession();
-    } else {
-      print('No valid session found, redirecting to RoleSelectionPage');
-      // No session at all, redirect to role selection
-      if (mounted) {
-        Navigator.pushReplacement(
-          context,
-          MaterialPageRoute(builder: (context) => const RoleSelectionPage()),
-        );
-      }
-    }
-
-    if (mounted) {
-      setState(() => _isLoading = false);
-    }
-  }
-
-  Future<void> _handleAuthenticatedSession(Session session) async {
-    print('User authenticated: ${session.user.id}');
-    print('Session expires at: ${session.expiresAt}');
-    print('Current time: ${DateTime.now()}');
-
-    // Check if session is expired (expiresAt is a timestamp in seconds)
-    if (session.expiresAt != null) {
-      final expiresAt = DateTime.fromMillisecondsSinceEpoch(
-        session.expiresAt! * 1000,
-      );
-      print('Session expires at: $expiresAt');
-
-      if (expiresAt.isBefore(DateTime.now())) {
-        print('Session expired, refreshing...');
-        try {
-          await _supabase.auth.refreshSession();
-          print('Session refreshed successfully');
-        } catch (e) {
-          print('Error refreshing session: $e');
-          await SessionService.clearSession();
-          return;
-        }
-      }
-    }
-
-    // Save session to persistent storage
+  Future<void> _checkAndRoute() async {
     try {
+      // Wait for Supabase auth state to be ready
+      await Future.delayed(const Duration(milliseconds: 500));
+
+      final session = _supabase.auth.currentSession;
+      if (session == null) {
+        // Not authenticated -> show role selection / login
+        _navigateToRoleSelection();
+        return;
+      }
+
+      final user = _supabase.auth.currentUser!;
+      // Fetch profile
       final profileResponse = await _supabase
           .from('profiles')
-          .select('role, phone')
-          .eq('id', session.user.id)
-          .single();
+          .select()
+          .eq('id', user.id);
 
-      final role = profileResponse['role'] as String;
-      final phoneNumber = profileResponse['phone'] as String? ?? '';
-      print('User role: $role, phone: $phoneNumber');
+      if (profileResponse.isEmpty) {
+        // No profile yet -> route to signup
+        _navigateToSignup();
+        return;
+      }
 
-      // Save session data
-      final expiry = session.expiresAt != null
-          ? DateTime.fromMillisecondsSinceEpoch(session.expiresAt! * 1000)
-          : DateTime.now().add(const Duration(days: 30));
+      final profile = profileResponse.first as Map<String, dynamic>;
+      final role = profile['role'] as String? ?? 'customer';
 
-      await SessionService.saveSession(
-        userId: session.user.id,
-        userPhone: phoneNumber,
-        userRole: role,
-        expiry: expiry,
-      );
-
-      // Redirect based on role
-      if (mounted) {
-        if (role == 'customer') {
-          print('Redirecting to CustomerHomePage');
-          Navigator.pushReplacement(
-            context,
-            MaterialPageRoute(builder: (context) => const CustomerHomePage()),
-          );
-        } else if (role == 'driver') {
-          print('Redirecting to DriverHomePage');
-          Navigator.pushReplacement(
-            context,
-            MaterialPageRoute(builder: (context) => const DriverHomePage()),
-          );
-        } else {
-          print('Invalid role, redirecting to RoleSelectionPage');
-          Navigator.pushReplacement(
-            context,
-            MaterialPageRoute(builder: (context) => const RoleSelectionPage()),
-          );
-        }
+      if (role == 'driver') {
+        await _handleDriverRouting(user.id, profile);
+      } else {
+        // Customer: go to customer home
+        _navigateToCustomerHome();
       }
     } catch (e) {
-      print('Error fetching profile or saving session: $e');
-      // Error fetching profile (likely no profile exists), redirect to role selection
+      debugPrint('Error in AuthWrapper routing: $e');
+      // Fallback to role selection on error
+      _navigateToRoleSelection();
+    } finally {
       if (mounted) {
-        CustomToast.showInfo(
-          context: context,
-          message: 'Please complete your profile setup',
-        );
-        Navigator.pushReplacement(
-          context,
-          MaterialPageRoute(builder: (context) => const RoleSelectionPage()),
-        );
+        setState(() => _isLoading = false);
       }
     }
   }
 
-  Future<void> _tryRestoreSession() async {
-    print('Attempting to restore session from local storage...');
-    final sessionData = await SessionService.getSessionData();
+  Future<void> _handleDriverRouting(
+    String userId,
+    Map<String, dynamic> profile,
+  ) async {
+    // Check if verification_status field exists and has a value
+    final verificationStatus = profile['verification_status'] as String?;
+    final isVerified = profile['is_verified'] as bool? ?? false;
 
-    if (sessionData != null) {
-      print('Found local session data, redirecting based on role');
-      final role = sessionData['userRole'] as String?;
+    // Fetch driver record
+    final driverResponse = await _supabase
+        .from('drivers')
+        .select()
+        .eq('id', userId);
 
-      if (mounted) {
-        if (role == 'customer') {
-          Navigator.pushReplacement(
-            context,
-            MaterialPageRoute(builder: (context) => const CustomerHomePage()),
-          );
-        } else if (role == 'driver') {
-          Navigator.pushReplacement(
-            context,
-            MaterialPageRoute(builder: (context) => const DriverHomePage()),
-          );
-        } else {
-          Navigator.pushReplacement(
-            context,
-            MaterialPageRoute(builder: (context) => const RoleSelectionPage()),
-          );
-        }
+    final driver = driverResponse.isNotEmpty
+        ? driverResponse.first as Map<String, dynamic>?
+        : null;
+    final vehicleType = driver != null
+        ? driver['vehicle_type'] as String?
+        : null;
+
+    debugPrint('Driver routing debug:');
+    debugPrint('  verificationStatus: $verificationStatus');
+    debugPrint('  isVerified: $isVerified');
+    debugPrint('  vehicleType: $vehicleType');
+
+    // Handle the initial state - if verification_status is null, it means the user
+    // hasn't started the verification process yet
+    if (verificationStatus == null) {
+      // New driver - route to vehicle type selection first
+      if (vehicleType == null || vehicleType.isEmpty) {
+        debugPrint('  Routing to vehicle type selection (new driver)');
+        _navigateToVehicleType(userId);
+      } else {
+        // Vehicle type is set but no verification status - route to verification
+        debugPrint(
+          '  Routing to verification (vehicle set but no verification)',
+        );
+        _navigateToVerification();
+      }
+    } else if (verificationStatus == 'pending') {
+      // If pending review, show waiting screen
+      debugPrint('  Routing to waiting review (pending verification)');
+      _navigateToWaitingReview();
+    } else if (verificationStatus == 'rejected') {
+      // If rejected, route to verification page to allow resubmission
+      debugPrint('  Routing to verification (rejected)');
+      _navigateToVerification();
+    } else if (verificationStatus == 'approved') {
+      // If approved but no vehicle_type, route to vehicle selection
+      if (vehicleType == null || vehicleType.isEmpty) {
+        debugPrint(
+          '  Routing to vehicle type selection (approved but no vehicle)',
+        );
+        _navigateToVehicleType(userId);
+      } else {
+        // All good -> route to enhanced driver home
+        debugPrint('  Routing to enhanced driver home (fully verified)');
+        _navigateToEnhancedDriverHome();
       }
     } else {
-      print('No valid local session found');
-      if (mounted) {
-        Navigator.pushReplacement(
-          context,
-          MaterialPageRoute(builder: (context) => const RoleSelectionPage()),
-        );
-      }
+      // Fallback to verification for unknown status
+      debugPrint(
+        '  Routing to verification (unknown status: $verificationStatus)',
+      );
+      _navigateToVerification();
     }
+  }
+
+  void _navigateToRoleSelection() {
+    Navigator.pushNamedAndRemoveUntil(
+      context,
+      '/role-selection',
+      (route) => false,
+    );
+  }
+
+  void _navigateToSignup() {
+    Navigator.pushNamedAndRemoveUntil(context, '/signup', (route) => false);
+  }
+
+  void _navigateToCustomerHome() {
+    Navigator.pushNamedAndRemoveUntil(
+      context,
+      '/customer-home',
+      (route) => false,
+    );
+  }
+
+  void _navigateToVehicleType(String driverId) {
+    Navigator.pushNamedAndRemoveUntil(
+      context,
+      '/vehicle-type-selection',
+      (route) => false,
+      arguments: driverId,
+    );
+  }
+
+  void _navigateToVerification() {
+    Navigator.pushNamedAndRemoveUntil(
+      context,
+      '/verification',
+      (route) => false,
+    );
+  }
+
+  void _navigateToWaitingReview() {
+    Navigator.pushNamedAndRemoveUntil(
+      context,
+      '/waiting-review',
+      (route) => false,
+    );
+  }
+
+  void _navigateToEnhancedDriverHome() {
+    Navigator.pushNamedAndRemoveUntil(
+      context,
+      '/enhanced-driver-home',
+      (route) => false,
+    );
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       body: Center(
-        child: _isLoading
-            ? const CircularProgressIndicator()
-            : const Text('Redirecting...'),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            const CircularProgressIndicator(),
+            const SizedBox(height: 16),
+            Text(
+              _isLoading ? 'Checking authentication...' : 'Redirecting...',
+              style: const TextStyle(fontSize: 16, color: Colors.grey),
+            ),
+          ],
+        ),
       ),
     );
   }
